# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EduVerify API
  description: >-
    Verification API for South African institutions and qualifications, backed by the
    eduverify-institutions DynamoDB table. Every route requires the x-api-key header except
    /v1/health, /v1/docs, and /v1/openapi.yaml. This file is generated by `npm run docs:generate`
    (src/docs/generateOpenApi.ts) from src/docs/routeSpec.ts and the TS types it references — do not
    hand-edit it.
  version: 0.0.1
servers:
  - url: https://{apiId}.execute-api.af-south-1.amazonaws.com/{stage}
    description: API Gateway stage (fill in apiId once infra is deployed)
    variables:
      apiId:
        default: your-api-id
      stage:
        default: staging
        enum:
          - staging
          - production
tags:
  - name: Health
  - name: Stats
  - name: Docs
  - name: Institutions
  - name: Qualifications
security:
  - ApiKeyAuth: []
paths:
  /v1/health:
    get:
      operationId: getHealth
      tags:
        - Health
      summary: Check API and DynamoDB health
      description: >-
        Unauthenticated. Returns "degraded" (not necessarily a 5xx) when the API process is up but
        its one dependency, the institutions table, isn't reachable.
      security: []
      responses:
        '200':
          description: Health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResult'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/stats:
    get:
      operationId: getStats
      tags:
        - Stats
      summary: Headline counts across the full corpus
      description: >-
        Scans every status partition (like /v1/institutions/list's status=ALL), not just REGISTERED
        — a total-corpus count rather than one filtered to a single register status.
      responses:
        '200':
          description: Corpus-wide counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResult'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/docs:
    get:
      operationId: getDocs
      tags:
        - Docs
      summary: Interactive Swagger UI for this spec
      description: >-
        Serves this same spec's Swagger UI page directly from the running API. The page fetches this
        spec from /v1/openapi.yaml.
      security: []
      responses:
        '200':
          description: Swagger UI HTML page
          content:
            text/html: {}
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/openapi.yaml:
    get:
      operationId: getOpenApiYaml
      tags:
        - Docs
      summary: This spec, as raw YAML
      description: The exact contents of this file, served for /v1/docs's Swagger UI to fetch.
      security: []
      responses:
        '200':
          description: OpenAPI 3 spec (YAML)
          content:
            text/yaml: {}
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/institutions/search:
    get:
      operationId: searchInstitutions
      tags:
        - Institutions
      summary: Fuzzy search institutions by name
      description: >-
        Scans every status partition and fuzzy-ranks the whole set (typo/word-order/abbreviation
        tolerant). An empty or missing q returns an empty, page-1 result rather than an error.
      parameters:
        - name: q
          in: query
          description: Free-text institution name query, e.g. "UCT" or "cape town".
          schema:
            type: string
        - name: province
          in: query
          description: Exact-match province filter.
          schema:
            type: string
          example: Western Cape
        - name: type
          in: query
          description: Exact-match institution type filter.
          schema:
            $ref: '#/components/schemas/InstitutionType'
        - name: page
          in: query
          description: 1-indexed page number.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 25
      responses:
        '200':
          description: One page of ranked search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/institutions/list:
    get:
      operationId: listInstitutions
      tags:
        - Institutions
      summary: Paginated browse/listing of institutions
      description: >-
        Fetches a full GSI1 status partition and paginates/filters in memory. status=ALL scans every
        partition instead of the default single status partition.
      parameters:
        - name: page
          in: query
          description: 1-indexed page number.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 25
        - name: province
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/InstitutionType'
        - name: status
          in: query
          description: >-
            GSI1PK status partition to read, or "ALL" to scan and merge every partition. See
            STATUS_PARTITIONS in src/lib/dynamodb.ts for the exhaustive list of valid values.
          schema:
            type: string
            default: REGISTERED
          example: REGISTERED
      responses:
        '200':
          description: One page of institutions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResult'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/institutions/{id}:
    get:
      operationId: getInstitution
      tags:
        - Institutions
      summary: Fetch a single institution by id
      description: Direct PK GetItem — finds an institution even if it isn't visible to GSI1-based search/list.
      parameters:
        - name: id
          in: path
          required: true
          description: The institution's DynamoDB PK, e.g. "INST#2000/HE07/015".
          schema:
            type: string
      responses:
        '200':
          description: The institution
          content:
            application/json:
              schema:
                type: object
                required:
                  - institution
                properties:
                  institution:
                    $ref: '#/components/schemas/InstitutionRecord'
        '404':
          description: No institution with that id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/institutions/verify:
    post:
      operationId: verifyInstitution
      tags:
        - Institutions
      summary: High-confidence identity verification for an institution
      description: >-
        The loan-org use case: a boolean-ish, high-confidence answer rather than a ranked list. A
        registration number that resolves is always "exact"; a name match is "exact" only when it
        equals a candidate's name after normalization, "high" otherwise.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyInstitutionRequest'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyInstitutionResult'
        '400':
          description: Malformed JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/qualifications/verify:
    post:
      operationId: verifyQualification
      tags:
        - Qualifications
      summary: Verify a single claimed qualification
      description: >-
        Given a CV's claimed (qualification title, institution name) pair, finds the institution and
        checks whether it actually offers that qualification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyQualificationRequest'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyQualificationResult'
        '400':
          description: Malformed JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/qualifications/verify/batch:
    post:
      operationId: verifyQualificationBatch
      tags:
        - Qualifications
      summary: Verify several claimed qualifications in one call
      description: >-
        One CV commonly claims several qualifications from the same institution, so this fans out to
        the single-item matcher per claim. Requires a tier with batch access (see src/tiers.ts);
        maxBatchSize is enforced per the caller's tier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/VerifyQualificationRequest'
      responses:
        '200':
          description: One verification result per input item, in the same order
          content:
            application/json:
              schema:
                type: object
                required:
                  - results
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/VerifyQualificationResult'
        '400':
          description: Malformed JSON body, or batch size exceeds the tier's maxBatchSize
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: This tier has no batch access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
      additionalProperties: false
      description: >-
        Wire shape of every non-2xx JSON body this router returns — see docs/openapi.yaml, generated
        from this type via src/docs/generateOpenApi.ts.
    HealthResult:
      type: object
      properties:
        status:
          type: string
          enum:
            - ok
            - degraded
        dynamodb:
          type: boolean
      required:
        - status
        - dynamodb
      additionalProperties: false
    StatsResult:
      type: object
      properties:
        totalInstitutions:
          type: number
        totalQualifications:
          type: number
        totalProvinces:
          type: number
      required:
        - totalInstitutions
        - totalQualifications
        - totalProvinces
      additionalProperties: false
    SearchResult:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/InstitutionRecord'
        page:
          type: number
        pageSize:
          type: number
        total:
          type: number
      required:
        - query
        - results
        - page
        - pageSize
        - total
      additionalProperties: false
    InstitutionRecord:
      type: object
      properties:
        name:
          type: string
        tradingName:
          type:
            - string
            - 'null'
        abbreviation:
          type:
            - string
            - 'null'
          description: Common short form, e.g. "UCT" — search-only, never a valid getDisplayName result.
        registration_number:
          type:
            - string
            - 'null'
        status:
          type:
            - string
            - 'null'
        address:
          type: string
        province:
          type:
            - string
            - 'null'
        contacts:
          $ref: '#/components/schemas/Contacts'
        faculties_and_programmes:
          type: array
          items:
            $ref: '#/components/schemas/FacultyProgrammes'
        cancellation_reason:
          type:
            - string
            - 'null'
        institutionType:
          $ref: '#/components/schemas/InstitutionType'
        isFeatured:
          type: boolean
        isSponsored:
          type: boolean
        isRecentlyAdded:
          type: boolean
        id:
          type: string
      required:
        - address
        - contacts
        - faculties_and_programmes
        - id
        - institutionType
        - name
      additionalProperties: false
    Contacts:
      type: object
      properties:
        email:
          type: array
          items:
            type: string
        phone:
          type: array
          items:
            type: string
        website:
          type:
            - string
            - 'null'
      required:
        - email
        - phone
      additionalProperties: false
    FacultyProgrammes:
      type: object
      properties:
        faculty:
          type: string
        programmes:
          type: array
          items:
            $ref: '#/components/schemas/SaqaQualification'
      required:
        - faculty
        - programmes
      additionalProperties: false
      description: >-
        An institution's SAQA-matched qualifications, grouped by subfield ("faculty"). Baked into
        data/institutions.json / public_universities.json / public_tvets.json by
        web/scripts/bakeFacultiesAndProgrammes.ts (in the eduverify repo) — an institution/faculty
        with no SAQA match gets an empty array, never omitted or null.
    SaqaQualification:
      type: object
      properties:
        qualId:
          type: number
        title:
          type: string
        nqfLevel:
          type: number
        nqfLevelRaw:
          type: string
        credits:
          type: number
        subfield:
          type: string
        originator:
          type: string
        framework:
          type: string
      required:
        - qualId
        - title
        - nqfLevelRaw
        - subfield
        - originator
        - framework
      additionalProperties: false
      description: >-
        A single SAQA NLRD qualification registration, from any NQF sub-framework (HEQSF, OQSF,
        GFETQSF, SFAP, SFNA). `framework` is the per-row discriminator for callers that need to
        filter to one sub-framework — e.g. EduVerify's own product is HEQSF-only, but this API's
        other consumers (CV/qualification verification) need the full set.
    InstitutionType:
      type: string
      enum:
        - Public University
        - Private Higher Education Institution
        - TVET College
    ListResult:
      type: object
      properties:
        institutions:
          type: array
          items:
            $ref: '#/components/schemas/InstitutionRecord'
        page:
          type: number
        pageSize:
          type: number
        total:
          type: number
      required:
        - institutions
        - page
        - pageSize
        - total
      additionalProperties: false
    VerifyInstitutionRequest:
      type: object
      properties:
        name:
          type: string
        registrationNumber:
          type: string
      additionalProperties: false
    VerifyInstitutionResult:
      type: object
      properties:
        matched:
          type: boolean
        confidence:
          type: string
          enum:
            - exact
            - high
            - none
        institution:
          $ref: '#/components/schemas/InstitutionRecord'
        status:
          type: string
      required:
        - matched
        - confidence
      additionalProperties: false
    VerifyQualificationRequest:
      type: object
      properties:
        qualificationTitle:
          type: string
        institutionName:
          type: string
        framework:
          type: string
          description: >-
            Optional NQF sub-framework filter (HEQSF, OQSF, GFETQSF, SFAP, SFNA). Omitted by default
            — unlike EduVerify's own product, this API is not HEQSF-only, since the CV/HR use case
            needs occupational and other qualifications verified too.
      required:
        - qualificationTitle
        - institutionName
      additionalProperties: false
    VerifyQualificationResult:
      type: object
      properties:
        matched:
          type: boolean
        confidence:
          type: string
          enum:
            - exact
            - fuzzy
            - none
        institution:
          $ref: '#/components/schemas/InstitutionRecord'
        qualification:
          $ref: '#/components/schemas/SaqaQualification'
      required:
        - matched
        - confidence
      additionalProperties: false
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key